+++ Frends.SFTP.DownloadFiles/CHANGELOG.md
@@ -3,9 +3,9 @@
 ## [2.7.0] - 2023-06-07
 ### Fixed
 - Modified private key passphrase to be visible when all private key authentication options were enabled.
-- Added new parameter for keyboard-interactive auhentication were users can add prompts and responses.
+- Added new parameter for keyboard-interactive authentication where users can add prompts and responses.
 - Modified operations log to list current system and sftp server information.
-- Fixed operations log to show case exceptions more precise.
+- Fixed operations log to show case exceptions more precisely.
 
 ## [2.6.1] - 2023-05-17
 ### Fixed